refactor: remove group backend service
